requirements.txt
asgiref==3.5.0
black==22.1.0
click==8.0.3
Django==3.1.7
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==2.4.0
django-leaflet==0.27.1
djangorestframework==3.12.2
djangorestframework-gis==0.17
flake8==4.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2021.3
sqlparse==0.4.2
tomli==2.0.0
typing_extensions==4.0.1
pytest-django==4.1.0
pytest-factoryboy==2.1.0
faker==8.1.1
pytest-cov==2.11.1
